refactor: 카테고리에서 슬라이더 분리 (#28)
