+++ FE/fe-sidedish/src/components/category/Category.style.js
@@ -1,5 +1,4 @@
 import styled from 'styled-components';
-import { ButtonType } from './Category';
 
 const Container = styled.section`
   margin-top: 50px;
@@ -19,25 +18,6 @@ const Title = styled.h2`
   margin-bottom: 40px;
 `;
 
-const Slide = styled.div`
-  position: relative;
-`;
-
-const SlideWrapper = styled.div`
-  overflow: hidden;
-`;
-
-const SlideButton = styled.button`
-  position: absolute;
-  top: 50%;
-  ${({ type }) => type === ButtonType.PREV && 'left: -60px;'}
-  ${({ type }) => type === ButtonType.NEXT && 'right: -60px;'}
-  width: 32px;
-  height: 32px;
-  font-size: 0;
-  transform: translateY(-50%);
-`;
-
 const CardList = styled.div`
   display: flex;
   justify-content: space-between;
@@ -48,8 +28,5 @@ export {
   Container,
   Wrapper,
   Title,
-  Slide,
-  SlideWrapper,
-  SlideButton,
   CardList,
 };
